refactor(api-gateway): move flat src into api_gateway package
Fourth app in the namespace refactor (second deployed one). api-gateway had
flat src/ modules (config, health, logging_config, main, security, telemetry)
that collide with other services' top-level modules when installed editable.

- Move all modules into src/api_gateway/ (git-tracked renames).
- Rewrite intra-app imports to relative (from .security import ...); tests
  import from api_gateway.<module>.
- Dockerfile: CMD uvicorn src.main:app -> uvicorn api_gateway.main:app
  (PYTHONPATH=.../src unchanged so api_gateway resolves).
- Makefile run-gateway: uvicorn main:app -> uvicorn api_gateway.main:app
  (keeps --app-dir api-gateway/src). No build config needed — runs from
  source (--no-install-project), not as an installed package.

Verified: api-gateway test passes; mypy clean (7 files); make lint + make
test green; uvicorn loads api_gateway.main:app (12 routes) locally and inside
a freshly built Docker image — the container's uvicorn "Started server
process" with the new CMD (subsequent lifespan error is an unrelated
connection-config parse on the dummy env, not the refactor).

Follow-up: core (last).
