+++ api-gateway/src/api_gateway/logging_config.py
