+++ api-gateway/src/api_gateway/telemetry.py
